+++ src/reset_grid_aabb.rs
@@ -20,10 +20,10 @@ impl ResetGridAABB {
     pub fn new(
         device: &Device,
         first_aabb: TypedBuffer<AABB>,
-        grid_min_x: TypedBuffer<i32>,
-        grid_min_y: TypedBuffer<i32>,
-        grid_max_x: TypedBuffer<i32>,
-        grid_max_y: TypedBuffer<i32>,
+        grid_min_x: TypedBuffer<f32>,
+        grid_min_y: TypedBuffer<f32>,
+        grid_max_x: TypedBuffer<f32>,
+        grid_max_y: TypedBuffer<f32>,
     ) -> Self {
         let bind_group = WgpuBindGroup0::from_bindings(
             device,
